--- server/src/modes/template/services/htmlHover.ts
@@ -1,93 +1,0 @@
-import { HTMLDocument } from '../parser/htmlParser';
-import { TokenType, createScanner } from '../parser/htmlScanner';
-import { TextDocument, Range, Position, Hover, MarkedString } from 'vscode-languageserver-types';
-import { IHTMLTagProvider } from '../tagProviders';
-import { NULL_HOVER } from '../../nullMode';
-
-const TRIVIAL_TOKEN = [TokenType.StartTagOpen, TokenType.EndTagOpen, TokenType.Whitespace];
-
-export function doHover(
-  document: TextDocument,
-  position: Position,
-  htmlDocument: HTMLDocument,
-  tagProviders: IHTMLTagProvider[]
-): Hover {
-  const offset = document.offsetAt(position);
-  const node = htmlDocument.findNodeAt(offset);
-  if (!node || !node.tag) {
-    return NULL_HOVER;
-  }
-  function getTagHover(tag: string, range: Range, open: boolean): Hover {
-    for (const provider of tagProviders) {
-      let hover: Hover | null = null;
-      provider.collectTags((t, label) => {
-        if (t === tag) {
-          const tagLabel = open ? '<' + tag + '>' : '</' + tag + '>';
-          hover = { contents: [{ language: 'html', value: tagLabel }, MarkedString.fromPlainText(label)], range };
-        }
-      });
-      if (hover) {
-        return hover;
-      }
-    }
-    return NULL_HOVER;
-  }
-
-  function getAttributeHover(tag: string, attribute: string, range: Range): Hover {
-    let hover: Hover = NULL_HOVER;
-    for (const provider of tagProviders) {
-      provider.collectAttributes(tag, (attr, _, documentation) => {
-        if (attribute !== attr) {
-          return;
-        }
-        const contents = [documentation ? MarkedString.fromPlainText(documentation) : `No doc for ${attr}`];
-        hover = { contents, range };
-      });
-    }
-    return hover;
-  }
-
-  const inEndTag = node.endTagStart && offset >= node.endTagStart; // <html></ht|ml>
-  const startOffset = inEndTag ? node.endTagStart : node.start;
-  const scanner = createScanner(document.getText(), startOffset);
-  let token = scanner.scan();
-
-  function shouldAdvance() {
-    if (token === TokenType.EOS) {
-      return false;
-    }
-    const tokenEnd = scanner.getTokenEnd();
-    if (tokenEnd < offset) {
-      return true;
-    }
-
-    if (tokenEnd === offset) {
-      return TRIVIAL_TOKEN.indexOf(token) > -1;
-    }
-    return false;
-  }
-
-  while (shouldAdvance()) {
-    token = scanner.scan();
-  }
-
-  if (offset > scanner.getTokenEnd()) {
-    return NULL_HOVER;
-  }
-  const tagRange = {
-    start: document.positionAt(scanner.getTokenOffset()),
-    end: document.positionAt(scanner.getTokenEnd())
-  };
-  switch (token) {
-    case TokenType.StartTag:
-      return getTagHover(node.tag, tagRange, true);
-    case TokenType.EndTag:
-      return getTagHover(node.tag, tagRange, false);
-    case TokenType.AttributeName:
-      // TODO: treat : as special bind
-      const attribute = scanner.getTokenText().replace(/^:/, '');
-      return getAttributeHover(node.tag, attribute, tagRange);
-  }
-
-  return NULL_HOVER;
-}
