--- server/src/types.ts
@@ -1,3 +1,0 @@
-export interface DocumentContext {
-  resolveReference(ref: string, base?: string): string;
-}
