--- server/src/utils/paths.ts
@@ -1,16 +1,0 @@
-import { platform } from 'os';
-import Uri from 'vscode-uri';
-
-export function getFileFsPath(documentUri: string): string {
-  return Uri.parse(documentUri).fsPath;
-}
-
-export function getFilePath(documentUri: string): string {
-  const IS_WINDOWS = platform() === 'win32';
-  if (IS_WINDOWS) {
-    // Windows have a leading slash like /C:/Users/pine
-    return Uri.parse(documentUri).path.slice(1);
-  } else {
-    return Uri.parse(documentUri).path;
-  }
-}
